qemu_domain: simplify non-VFIO memLockLimit calc for PPC64
passthroughLimit is being calculated even if usesVFIO is false.
After that, a if/else conditional is used to check if we're going
to sum it up with baseLimit.

This patch initializes passthroughLimit to zero and always
return memKB = baseLimit + passthroughLimit. The conditional
is then used to calculate passthroughLimit if usesVFIO is true.
This results in some cycles spared for the usesVFIO=false
scenario, but the real motivation is to make the code simpler
to add an alternative passthroughLimit formula for NVLink2
passthrough.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>
Message-Id: <20190303132314.27814-2-danielhb413@gmail.com>