qemu_domain: NVLink2 device tree functions for PPC64
The NVLink2 support in QEMU implements the detection of NVLink2
capable devices by verfying the attributes of the VFIO mem region
QEMU allocates for the NVIDIA GPUs. To properly allocate an
adequate amount of memLock, Libvirt needs this information before
a QEMU instance is even created.

An alternative is presented in this patch. Given a PCI device,
we'll traverse the device tree at /proc/device-tree to check if
the device has a NPU bridge, retrieve the node of the NVLink2 bus,
find the memory-node that is related to the bus and see if it's a
NVLink2 bus by inspecting its 'reg' value. This logic is contained
inside the 'device_is_nvlink2_capable' function, which uses other
new helper functions to navigate and fetch values from the device
tree nodes.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>
Message-Id: <20190303132314.27814-3-danielhb413@gmail.com>