PPC64 support for NVIDIA V100 GPU with NVLink2 passthrough
The NVIDIA V100 GPU has an onboard RAM that is mapped into the
host memory and accessible as normal RAM via an NVLink2 bus. When
passed through in a guest, QEMU puts the NVIDIA RAM window in a
non-contiguous area, above the PCI MMIO area that starts at 32TiB.
This means that the NVIDIA RAM window starts at 64TiB and go all the
way to 128TiB.

This means that the guest might request a 64-bit window, for each PCI
Host Bridge, that goes all the way to 128TiB. However, the NVIDIA RAM
window isn't counted as regular RAM, thus this window is considered
only for the allocation of the Translation and Control Entry (TCE).

This memory layout differs from the existing VFIO case, requiring its
own formula. This patch changes the PPC64 code of
qemuDomainGetMemLockLimitBytes to:

- detect if a VFIO PCI device is using NVLink2 capabilities. This is
done by using the device tree inspection mechanisms that were
implemented in the previous patch;

- if any device is a NVIDIA GPU using a NVLink2 bus, passthroughLimit
is calculated in a different way to account for the extra memory
the TCE table can alloc. The 64TiB..128TiB window is more than
enough to fit all possible GPUs, thus the memLimit is the
same regardless of passing through 1 or multiple V100 GPUs.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>
Message-Id: <20190303132314.27814-4-danielhb413@gmail.com>