libxl: Add support for max_grant_frame setting
Xen 4.10 introduced the max_grant_frames xl config setting, which
can be set globally in xl.conf(5) or per-domain in xl.cfg(5).
max_grant_frames specifies the maximum number of grant frames the
domain is allowed to have, which in turn controls the number of
pages the domain can share.

This patch adds support for setting max_grant_frames on a global
level in libxl.conf. Per-domain support via domXML can be provided
in a future patch.

Signed-off-by: Jim Fehlig <jfehlig@suse.com>
Message-Id: <20190304230001.18690-1-jfehlig@suse.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>