lxc: Rebase lxcNetworkParseData struct to support indexes.
The current logic of lxcNetworkParseData uses one single structure to
record data over the network definitions inside config files. The logic
consider the entry 'type' as a new network definition, every time that
algorithm find this tag.

This new structure was designed to consider network definitions as an
array of network structures.

Signed-off-by: Julio Faracco <jcfaracco@gmail.com>
Message-Id: <20190305015432.8537-2-jcfaracco@gmail.com>