domain: Add struct for future domain format parameters
Upcoming patches will add new flags for increasing the amount of
information present in <domain> dumpxml, but where the source
of that information is tied to somewhere other than the active
or offline domain sub-definition.  Make those extensions easier
by updating internal callers to pass in a struct, rather than
adding new parameters for each extension, so that later patches
only have to patch callers that care about a new member of the
struct rather than all callers.

I considered updating virDomainDefFormat(), but as there were
so many existing callers, it was easier to just add a new
wrapper function virDomainDefFormatFull() which takes the full
struct, while making the existing interface forward on to the
full one.

Since all callers are being adjusted anyway, reorder the
parameters of virDomainDefFormatInternal to put buf first, as
that tends to be more typical.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20190305033445.17140-8-eblake@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>