qemu: Implement bulk snapshot operations
Implement the new flags for bulk snapshot dump and redefine. This
borrows from ideas in the test driver, but is further complicated
by the fact that qemu must write snapshot XML to disk, and thus must
do additional validation after the initial parse to ensure the user
didn't attempt to rename a snapshot with "../" or similar.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20190305033445.17140-19-eblake@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>