conf: Accept model=virtio for SCSI controllers
At the moment, all VirtIO devices support model=virtio except
for SCSI controllers where model=virtio-scsi must be used
instead: get rid of this inconsistency by providing an alias
at the parser level, so that existing code keeps working but
using the same values across the board is also optionally
possible.

Tweak one of the test cases to show that the new value is
accepted transparently, without altering any of the output
files in the process.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Message-Id: <20190305150740.27408-3-abologna@redhat.com>