tests: add targets for building libvirt inside docker containers
The Travis CI system uses docker containers for its build environment.
These are pre-built and hosted under quay.io/libvirt so that developers
can use them for reproducing problems locally.

Getting the right docker command syntax to use them, however, is not
entirely easy. This patch addresses that usability issue by introducing
some make targets. To run a simple build (aka 'make all') using the
Fedora 28 containr:

   make cibuild-fedora-28

To also run unit tests

   make cicheck-fedora-28

This is just syntax sugar for calling the previous command with a
custom make target

   make cibuild-fedora-28 MAKE_ARGS="check"

To do a purely interactive build it is possible to request a shell

   make cishell-fedora-28

To do a mingw build, it is currently possible to use the fedora-rawhide
and request a different configure script

   make cibuild-fedora-rawhide CONFIGURE=mingw32-configure

In all cases the GIT source tree is cloned locally into a 'citree/src'
sub-directory which is then exposed to the container at '/build'. It is
setup to facilitate VPATH build so the initial working directory
is '/build/vpath'. An in source tree build can be requested instead
by passing  VPATHDIR= SRCDIR=.  args to make.

The make rules are kept in a standalone file that is included into the
main Makefile.am, so that it is possible to run them without having to
invoke autotools first.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>
Message-Id: <20190306093424.22469-2-berrange@redhat.com>