qemu: command: Override HOME variable for QEMU processes
By default, qemu user's home dir points to '/' which shouldn't be used
at all. We therefore override the HOME variable and pass the current
environment setting iff not running as SUID, which means that for
systemd we never set the variable. To enhance isolation of QEMU
processes, we should always override the HOME variable and point it
somewhere meaningful, like 'libDir/<domain>'.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <45600eda56353cc78c4a6a48177cd621ba8fcaae.1551882194.git.eskultet@redhat.com>