conf: domcaps: Add virCapsEnum 'report'
virCapsEnum report is an internal bool indicating whether we
should format the enum in the XML at all. This is unused for
now but will be handled in future patches.

We use a plain bool instead of tristate because the case here
is a bit different than the explicit @supported output. We
already report the equivalent of supported=YES|NO based on
what enum values are filled in. This adds report=false to
handle the ABSENT case.

Signed-off-by: Cole Robinson <crobinso@redhat.com>
Message-Id: <625eca459817ea253c1b2a0f60961935b56b9e13.1551914794.git.crobinso@redhat.com>