qemu: command: Override HOME variable for system QEMU
By default, qemu user's home dir points to '/' which shouldn't be used
at all. We therefore pass the HOME variable from the current variable
iff not running as SUID, which means that for systemd we never set it.
This patch makes sure, that for system QEMU this is always set to
libDir/<driver>, session mode is left untouched.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <f4b033010f91c198696a2d9c27a9409b82ea6c43.1552045608.git.eskultet@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>