rbd: optionally disable actual disk-usage during pool/volume refresh
For pools with numerous volumes or very large volumes, the disk-usage
calculation can take a non-trivial amount of time even with the
fast-diff feature enabled (especially since the the usage is calculated
serially for each image in the pool).

The "rbd:config_opts" node now supports a new libvirt-internal
"libvirt_calculate_disk_usage" option which can be set to false to
disable this calculation as needed.

Signed-off-by: Jason Dillaman <dillaman@redhat.com>
Message-Id: <20190311153955.4131-3-dillaman@redhat.com>