lxc: Introduce method lxcDomainGetBlockStats().
This method is responsible to fetch all Block Stats and store data into
virDomainStatsRecordPtr structure. This particular method checks for LXC
filesystem in host disk/partition and check for block disks like
physical disks and partition that are considered block types. If the
block type does not have a CGroup associated, it returns an error.

Signed-off-by: Julio Faracco <jcfaracco@gmail.com>
Message-Id: <20190311205618.6588-6-jcfaracco@gmail.com>