snapshot: Add new API for bulk dumpxml/redefine
Right now, copying the state of a transient domain with snapshots from
one host to another requires multiple API calls on both machines - on
the host: get the domain XML, get a list of the snapshots, and then
for each snapshot get the snapshot's XML; then on the destination:
create the domain, then multiple domain snapshot create calls (must
be in topological order) with the REDEFINE flag. This patch aims to
make the process use fewer APIs by making it possible to grab the XML
for all snapshots at once on the source, then redefine from that same
list at the destination. In fact, although it is easy to match recent
changes to allow output in topological order, it is also easy to do
bulk redefine in arbitrary order.

Consideration was given to adding a flag to existing APIs; but to do
that, there would be an asymmetry of dumping <snapshots> as a
sub-element of <domain> during virDomainGetXMLDesc(), vs. redefining
<snapshots> as a top-level element (and different than the normal
top-level element) of virDomainSnapshotCreateXML(). Thus, it is
cleaner to have two new APIs:

virDomainGetSnapshotsXMLDesc
virDomainImportSnapshotsXML

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20190312023839.11069-2-eblake@redhat.com>
Reviewed-by: Ján Tomko <jtomko@redhat.com>
Based-on: <20190308060512.17733-1-eblake@redhat.com>