snapshot: Tweaks to support new bulk dumpxml/import API
Change the return value of virDomainSnapshotObjLisParse() to return
the number of snapshots imported, and allow a return of 0 (the
original proposal of adding a flag to virDomainSnapshotCreateXML
required returning an arbitrary non-NULL snapshot, but with a new API
that returns a count, we are no longer constrained to a non-empty
list).

Change virDomainSnapshotObjListFormat()'s flags argument to be
the new public virDomainGetSnapshotsXMLFlags, since it is easy
to support both flag values.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20190312023839.11069-4-eblake@redhat.com>
Reviewed-by: Ján Tomko <jtomko@redhat.com>
Based-on: <20190308060512.17733-1-eblake@redhat.com>