virsh: Expose bulk snapshot dumpxml/import
Add a new 'snapshot-import' command for bulk import. For bulk
XML, it was easier to just add new flags to the existing
'snapshot-dumpxml' than to figure out a new command name.

I debated about whether omitting the snapshotname should be
enough to trigger the new API call, or whether to require a
new bool --all; in the end, I went with the latter.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20190312023839.11069-6-eblake@redhat.com>
Reviewed-by: Ján Tomko <jtomko@redhat.com>
Based-on: <20190308060512.17733-1-eblake@redhat.com>