test: Implement bulk snapshot operations
Implement the new API calls for bulk snapshot dump and import. The
bulk of the work is already done by the common code.

Since each connection to test:///default restarts at the same
canned state, this can easily be tested with:

$ virsh -c test:///default "
   snapshot-create-as test s1
   snapshot-create-as test s2
   echo witness
   snapshot-dumpxml test --all" | sed '1,/witness/d' > list.xml
$ virsh -c test:///default "
   snapshot-list test
   snapshot-import test list.xml
   snapshot-current --name test
   snapshot-list --parent test
"
 Name   Creation Time   State
-------------------------------

Imported 2 snapshots
s2
 Name   Creation Time               State     Parent
------------------------------------------------------
 s1     2019-02-20 22:26:52 -0600   running
 s2     2019-02-20 22:26:52 -0600   running   s1

The test driver also makes it easy to test input validation, by
modifying list.xml incorrectly (such as trying to attempt circular
dependencies).  Proving that that --topological makes a difference
is a bit harder (since we randomize the hash seed, it is not 100%
reproducible which order you get without it), but I found that
creating snapshots s1, s3, s2 in that order tended to be more likely
to hash in non-topological order without the flag.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20190312023839.11069-7-eblake@redhat.com>
Reviewed-by: Ján Tomko <jtomko@redhat.com>
Based-on: <20190308060512.17733-1-eblake@redhat.com>