qemu: Implement bulk snapshot operations
Implement the new API calls for bulk snapshot dump and import. This
borrows from ideas in the test driver, but import is further
complicated by the fact that qemu writes snapshot XML to disk, and
thus must do additional validation after the initial parse to ensure
the user didn't attempt to create a snapshot with "../" or similar.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20190312023839.11069-9-eblake@redhat.com>
Reviewed-by: Ján Tomko <jtomko@redhat.com>
Based-on: <20190308060512.17733-1-eblake@redhat.com>