tools: console: cleanup console on errors in main thread
We only check now for virObjectWait failures in virshRunConsole but
we'd better check and for other failures too. Anyway if failure
happened we need to shutdown console to stop delivering events
from event loop thread or we are in trouble as console is freed
on virshRunConsole exit.

We need to turn console into virObject object because stream/fd callbacks
can be called from a event loop thread after shutdown/freeing console
in main thread. It is convinient to turn into virLockableObject as
we have mutex in console object.

Signed-off-by: Nikolay Shirokovskiy <nshirokovskiy@virtuozzo.com>
Message-Id: <1552462789-715109-2-git-send-email-nshirokovskiy@virtuozzo.com>