test: Drop a few unnecessary <source> elements added by virt-install
The <source mode='bind'/> elements were added in commit 4d7ea75e and
apparently generated by virt-install, but since the @path attribute
is missing, the element will never be formatted back to the XML which the
libvirt-go-xml-check expects, so it fails.
The lack of the @path attribute is not a problem in general, because we
can autogenerate the path in certain cases, so qemuxml2argvtest will
always be happy, but XML formatting is a different thing.

Therefore, drop these generated elements, as they're unnecessary for the
purposes of qemuxml2argvtest and won't mess with the XML formatting
test in Go.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <1794865d03d6e51497bb5f487bd2524eec2027a0.1552466451.git.eskultet@redhat.com>