virt-yajl.m4: Drop useless check for qemu
Firstly, this code is lacking AC_REQUIRE() therefore it may
happen that this code is executed before
LIBVIRT_DRIVER_CHECK_QEMU is evaluated. Secondly, the code tries
to detect installed version of qemu to learn if it uses HMP or
QMP and enable YAJL based on that. Well, we support only QMP and
also minimal required version of qemu is 1.5.0 so the check would
have enabled yajl anyway.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <9c5edf231818ddfc03ff9cbcb0342b8238f2f37f.1552492158.git.mprivozn@redhat.com>
Reviewed-by: Ján Tomko <jtomko@redhat.com>