conf: net: Add model enum, and netfront value
This adds a network model enum. The virDomainNetDef property
is named 'model' like most other devices.

When the XML parser or a driver calls NetSetModelString, if
the passed string is in the enum, we will set net->model,
otherwise we copy the string into net->modelstr

Add a single example for the 'netfront' xen model, and wire
that up, just to verify it's all working

Signed-off-by: Cole Robinson <crobinso@redhat.com>
Message-Id: <ee6795bd0367efbc3d99718a578b8e755a75f512.1552492022.git.crobinso@redhat.com>