tests: qemuxml2argv: add va_arg enum handling
This establishes a pattern that will allow us to make test macros
more general purpose, by taking optional arguments. The general
format will be:

DO_TEST_FULL(...
             ARG_FOO, <value1>,
             ARG_BAR, <value2>)

ARG_X are just enum values that we look for in the va_args and know
how to interpret.

Implement this for the existing implicit qemuCaps va_args

Signed-off-by: Cole Robinson <crobinso@redhat.com>
Message-Id: <185092267f0b872f526bd5b9f17372ce252b53da.1552574299.git.crobinso@redhat.com>
Reviewed-by: Andrea Bolognani <abologna@redhat.com>