snapshot: Use accessors for virDomainSnapshot members
Upcoming patches want to add virDomainCheckpoint that behaves very
similarly to virDomainCheckpoint; the easiest way to share common code
is to give both classes a common base class. If this were C++, we'd
just use public member inheritance; but since it is C, we instead have
to touch EVERY use of member fields that will be relocated. To avoid
having to make future edits, wrap the majority of accesses behind
static inline functions, so that we only have to touch one place when
changing class inheritance; and temporarily rename the members to let
the compiler ensure we caught all uses.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20190315050233.10782-7-eblake@redhat.com>