backup: Introduce virDomainCheckpointPtr
Prepare for introducing a bunch of new public APIs related to backup
checkpoints by first introducing a new internal type and errors
associated with that type.  Checkpoints are modeled heavily after
virDomainSnapshotPtr (both represent a point in time of the guest),
although a snapshot exists with the intent of rolling back to that
state, while a checkpoint exists to make it possible to create an
incremental backup at a later time.  Thus, it shares the common
virDomainMoment base class created in the previous patches.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20190315050233.10782-9-eblake@redhat.com>