storage: split off code for calling rbd_list
The rbd_list method has a quite unpleasant signature returning an
array of strings in a single buffer instead of an array. It is
being deprecated in favour of rbd_list2. To maintain clarity of
code when supporting both APIs in parallel, split the rbd_list
code out into a separate method.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>
Message-Id: <20190318122018.9715-2-berrange@redhat.com>
Reviewed-by: Ján Tomko <jtomko@redhat.com>