snapshot: Use accessors for virDomainSnapshot members
Upcoming patches want to add virDomainCheckpoint that behaves very
similarly to virDomainSnapshot; the easiest way to share common code
is to give both classes a common base class. If this were C++, we'd
just use public member inheritance; but since it is C, we instead have
to touch EVERY use of member fields that will be relocated. To avoid
having to make future edits, wrap the majority of accesses behind
static inline functions, so that we only have to touch one place when
changing class inheritance.

I wrote the patch by temporarily renaming the members so the compiler
would force me to find all uses, but the final commit does not
preserve that rename.

Signed-off-by: Eric Blake <eblake@redhat.com>
Reviewed-by: Ján Tomko <jtomko@redhat.com>
Message-Id: <20190320054105.17689-3-eblake@redhat.com>