snapshot: Track current snapshot in virDomainSnapshotObjList
It is easier to track the current snapshot as part of the list of
snapshots. In particular, doing so lets us guarantee that the current
snapshot is cleared if that snapshot is removed from the list (rather
than depending on the caller to do so, and risking a use-after-free
problem).  This requires the addition of several new accessor
functions.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20190320054105.17689-7-eblake@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>